libselinux: move flush call to avoid race condition
Userspace policy enforcers like dbus-daemon need to update their cached
info when the policy gets reloaded. The selinux_netlink_policyload
callback can be configured by the users of libselinux to trigger
whatever other reload activity is necessary. In the case of the
dbus-daemon, a notification is sent to a file descriptor, so the main
event loop can handle the actual reloading.

The race condition now is that if another client connection gets
serviced between the call to selinux_flush_class_cache in
libselinux/src/avc_internal.c:L94 and the handling of the reload event
in the dbus-daemon main loop, that client will be rejected by the
selinux permission check, because the selinux object class "dbus" will
be unknown at that point.

In other words, the cache flush should be moved into the responsibility
of the policy reload callback, not done automatically, in order to give
callback implementers a chance to synchronize the cache flush with the
rest of their reloading activity.

To keep behavior consistent for users of libselinux that don't override
the policyreload callback, add the cache flush call to the default
callback implementation.

Signed-off-By: Johannes Schilling <johannes.schilling@partner.bmw.de>